refactor: move code to proper module to break cyclic dependency
